refactor: update Order class to use DECIMAL_ZERO for totalQuantity and filledQuantity, and adjust trade converter tests for Decimal handling

+++ ib_async/protobuf_converters/trade_converters.py
@@ -544,7 +544,7 @@ def createOrder(
     if orderProto.HasField("action"):
         order.action = orderProto.action
     if orderProto.HasField("totalQuantity"):
-        order.totalQuantity = float(orderProto.totalQuantity)
+        order.totalQuantity = Decimal(orderProto.totalQuantity)
     if orderProto.HasField("orderType"):
         order.orderType = orderProto.orderType
     if orderProto.HasField("lmtPrice"):
